dj_test/requirement.txt
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
decorator==4.0.9
Django==1.9.6
django-bootstrap3==7.0.1
django-ckeditor==5.0.3
ipython==4.2.0
ipython-genutils==0.1.0
Markdown==2.6.6
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
Pillow==3.2.0
ptyprocess==0.5.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
tornado==4.3
traitlets==4.2.1
uWSGI==2.0.12
